Restore block variations
